Move route building to conftest
